APPINT-32940

+++ main/plugins/org.talend.camel.designer/src/main/java/org/talend/camel/designer/build/RouteBundleExportAction.java
@@ -23,9 +23,9 @@ package org.talend.camel.designer.build;
 
 import java.io.File;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import org.eclipse.core.runtime.IProgressMonitor;
@@ -122,7 +122,7 @@ public class RouteBundleExportAction extends JobExportAction {
     @Override
     protected void doArchiveExport(IProgressMonitor monitor, List<ExportFileResource> resourcesToExport) {
 
-        Collection<String> unSelectedBundles = new ArrayList();
+    	Map<String, Boolean> unSelectedBundles = new HashMap<String, Boolean>();
 
         if (resourcesToExport.size() > 0) {
             FilesUtils.emptyFolder(getTemporaryStoreFile(new File(""), LIB));
@@ -159,13 +159,13 @@ public class RouteBundleExportAction extends JobExportAction {
                             }
 
                             boolean exist = false;
-                            for (String name : unSelectedBundles) {
+                            for (String name : unSelectedBundles.keySet()) {
                                 if (name.equals(file.getName())) {
                                     exist = true;
                                 }
                             }
 
-                            if (!exist) {
+                            if (!exist && unSelectedBundles.get(file.getName())) {//??????
                                 FilesUtils.copyFile(file, getTemporaryStoreFile(file, LIB));
                             }
